auth-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic[email]==2.9.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aio-pika==9.4.1
python-json-logger==3.1.0
email-validator==2.2.0
